# docs/api.yaml
openapi: 3.0.1
info:
  title: Plataforma de cursos do IFRO
  description: API responsável pelo controle da plataforma de cursos ofertados pelo IFRO
  termsOfService: ''
  contact: 
    email: 'email@email.com'
  license: 
    name: MIT
    url: https://opensourse.org/licenses/MIT
  version: 1.0.0
...
paths:
...
  /curso/{cursoid}/matricular:
    post:
      tags:
      - Usuário
      summary: Matricular em um curso
      description: Essa ação só pode ser realizada por usuários autenticados
      operationId: MatricularEmCurso
      parameters:
      - name: cursoid
        in: path
        description: O id do curso que será matriculado
        required: true
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: Usuário matriculado com sucesso
          content: {}
        400:
          description: Usuário já matriculado nesse curso 
          content: {}
        404:
          description: Curso não encontrado
          content: {}
...